refactor: enumerate 디렉토리 이름 enums로 변경
